requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django==1.11.5
djangorestframework-jwt==1.11.0
djangorestframework==3.6.4
pyjwt==1.5.3              # via djangorestframework-jwt
pytz==2017.2              # via django
